--- src/app/projects/activity-editor/project-add-activity.component.html
@@ -1,128 +1,0 @@
-<div class="editor-container editor-container-style">
-  <div class="header border-bottom">
-    <div class="close-button right-position" (click)="onClose()">&times;</div>
-    <br>
-    <div class="title ">Agregar actividad al proyecto: <b>{{project.name}}</b></div>
-  </div>
-  <br>
-  <div>
-    <table class="form">
-      <tr>
-        <td colspan="3" class="form-title form-title-text">1) Seleccionar una actividad o evento</td>
-      </tr>
-      <tr>
-        <td>Categoría:</td>
-        <td colspan="2"><select class="select-box" style="width:400px">
-          <option value="">( Todas )</option>          
-          <option value="Actividades de exploración">Período de transición de arranque</option>          
-          <option value="Actividades de exploración">Período de evaluación</option>
-          <option value="Actividades de explotación">Período de desarrollo</option>
-          <option value="Actividades de abandono">Etapa de abandono</option>
-          <option value="Actividades administrativas">Actividades administrativas</option>
-          <option value=""></option>
-          <option value="Actividades de exploración">Actividades de exploración</option>
-          <option value="Actividades de explotación">Actividades de explotación</option>
-          <option value="Actividades de mantenimiento">Mantenimiento</option>
-          <option value="Actividades de mantenimiento">Seguridad industrial</option>
-          <option value="Actividades petroleras">Otras actividades petroleras</option>          
-          <option value=""></option>          
-          <option value="Eventos">Eventos</option>
-        </select></td>
-      </tr>
-      <tr>
-        <td>Actividad:</td>
-        <td>
-          <select class="select-box" style="width:400px" [(ngModel)]="activity.name">
-            <option value="">( Todas )</option>
-            <option value="Ampliación de almacenamiento de residuos peligrosos">Ampliación de almacenamiento de residuos peligrosos</option>
-            <option value="Informar sobre caso fortuito">Informar sobre caso fortuito</option>
-            <option value="Certificación de reservas">Certificación de reservas</option>
-            <option value="Disposición final de residuos peligrosos">Disposición final de residuos peligrosos</option>       
-            <option value="Modificación de medidores">Modificación de medidores</option>
-            <option value="Proceso de unificación">Proceso de unificación</option>
-            <option value="Taponamiento de pozos">Taponamiento de pozos</option>
-          </select>
-        </td>
-        <td><a class="right-position">Actividades...</a></td>
-      </tr>
-      <tr>
-        <td>Asociado a:</td>
-        <td><select class="select-box" style="width:400px" [(ngModel)]="activity.resourceUID">
-          <option value="">( Seleccionar )</option>
-          <option *ngFor="let resource of resourceList" 
-                  [value]="resource.uid">{{resource.name}}</option>         
-          </select>
-        </td>
-        <td><a class="right-position">Instalaciones...</a></td>
-      </tr>
-      <tr>
-        <td colspan="3">
-          <table class="form internal-form">
-            <tr>
-              <td>Solicitado por:</td>
-              <td><select class="select-box" style="width:380px" [(ngModel)]="activity.requestedByUID">
-                <option value="">( Seleccionar )</option>
-                 <option *ngFor="let requester of requestersList" [value]="requester.uid">{{requester.name}}</option>                  
-                </select>
-
-              </td>
-              <td>El día:</td>
-              <td><input type="text" id="requestedDateCalendar" class="text-box" style="width:100px" [(ngModel)]="activity.requestedTime"></td>
-              <td><button class="btn " id="requestedDateButton" class="right-position">..</button></td>
-            </tr>
-          </table>
-
-        </td>
-      </tr>
-    </table>
-  </div>
-  <div>
-    <table class="form">
-      <tr>
-        <td colspan="3" class="form-title form-title-text">2) Información adicional</td>
-      </tr>
-      <tr>
-        <td>Responsable:</td>
-        <td><select class="select-box" style="width:440px" [(ngModel)]="activity.responsibleUID">
-                <option value="">( Seleccionar )</option>
-                <option *ngFor="let responsible of responsiblesList" 
-                        [value]="responsible.uid">{{responsible.name}}</option>            
-                </select>
-        </td>
-        <td><a class="right-position">Recursos...</a></td>
-      </tr>
-      <tr>
-        <td colspan="3">
-          <table class="form internal-form">
-            <tr>
-              <td class="text-align-left">Fecha de inicio<br>planeada:</td>
-              <td><input type="text" id="startDateCalendar" class="text-box" style="width:100px" [(ngModel)]="activity.estimatedStart"></td>
-              <td><button id="startDateCalendarButton" class="btn ">..</button></td>
-              <td>Fecha de terminación<br>estimada:</td>
-              <td><input id="endDateCalendar" type="text" class="text-box" style="width:100px" [(ngModel)]="activity.estimatedEnd"></td>
-              <td><button id="endDateCalendarButton" class="btn ">..</button></td>
-              <td><a class="right-position">73 días</a></td>
-            </tr>
-          </table>
-        </td>
-      </tr>
-
-      <tr>
-        <td>Observaciones:</td>
-        <td colspan="2"><textarea style="width:100%;" class="text-area" rows="5" cols="25" [(ngModel)]="activity.notes"></textarea></td>
-      </tr>
-    </table>
-  </div>
-  <div class="form-footer border-top">
-    <table class="form">
-      <tr>
-        <td class="text-align-right ">
-          <button class="btn " (click)="onClickCancel()">Cancelar</button>
-        </td>
-        <td class="text-align-right ">
-          <button class="btn " (click)="onClickAddActivity()">Agregar actividad</button>
-        </td>
-      </tr>
-    </table>
-  </div>
-</div>
