--- src/app/projects/activity-editor/project-start-activity.component.html
@@ -1,123 +1,0 @@
-<div class="editor-container editor-container-style">
-  <div class="header border-bottom">
-    <div class="close-button right-position" (click)="onClose()">&times;</div>
-    <br>
-    <div class="title ">Iniciar actividad:</div>
-  </div>
-
-  <div>
-    <table class="form">
-      <tr>
-        <td colspan="3" class="form-title form-title-text">1) Información de la actividad</td>
-      </tr>
-      <tr>
-        <td>Acitividad:</td>
-        <td><select [disabled]="disabled" class="select-box" style="width:400px" [(ngModel)]="activity.name">
-          <option value="">( Todas )</option>
-          <option value="Modificación de medidores">Modificación de medidores</option>         
-          </select>
-        </td>
-        <td></td>
-      </tr>
-      <tr>
-        <td>Asociado a:</td>
-        <td><select [disabled]="disabled" class="select-box" style="width:400px" [(ngModel)]="activity.resourceUID">
-          <option value="">( Seleccionar )</option>
-          <option *ngFor="let resource of resourceList" 
-                  [value]="resource.uid">{{resource.name}}</option>         
-          </select>
-        </td>
-        <td></td>
-      </tr>
-      <tr>
-        <td colspan="3">
-          <table class="internal-form">
-            <tr>
-              <td class="text-align-left">Solicitado por:</td>
-              <td>
-                <select [disabled]="disabled" class="select-box" style="width:380px" [(ngModel)]="activity.requestedByUID">
-                  <option value="">( Seleccionar )</option>
-                  <option *ngFor="let requester of requestersList" [value]="requester.uid">{{requester.name}}</option>
-                  </select>
-              </td>
-              <td>El día:</td>
-              <td><input [disabled]="disabled" type="text" id="requestedDateCalendar" class="text-box" style="width:100px" [(ngModel)]="activity.requestedTime"></td>
-              <td><button [disabled]="disabled" class="btn " id="requestedDateButton" class="right-position">..</button></td>
-            </tr>
-          </table>
-        </td>
-      </tr>
-      <tr>
-        <td>Responsable:</td>
-        <td><select [disabled]="disabled" class="select-box" style="width:440px" [(ngModel)]="activity.responsibleUID">
-                <option value="">( Seleccionar )</option>
-                <option *ngFor="let responsible of responsiblesList" 
-                        [value]="responsible.uid">{{responsible.name}}</option>            
-                </select>
-        </td>
-        <td></td>
-      </tr>
-      <tr>
-        <td colspan="3">
-          <table class="internal-form">
-            <tr>
-              <td class="text-align-left">Fecha de inicio planeada:</td>
-              <td><input [disabled]="disabled" type="text" id="startDateCalendar" class="text-box" style="width:100px" [(ngModel)]="activity.estimatedStart"></td>
-              <td>
-                <button [disabled]="disabled" id="startDateCalendarButton" class="btn ">..</button>
-              </td>
-              <td>Fecha de terminación estimada:</td>
-              <td><input [disabled]="disabled" id="endDateCalendar" type="text" class="text-box" style="width:100px" [(ngModel)]="activity.estimatedEnd"></td>
-              <td><button [disabled]="disabled" id="endDateCalendarButton" class="btn ">..</button></td>
-              <td><a class="right-position">{{activity.estimatedDuration}}</a></td>
-            </tr>
-          </table>
-        </td>
-      </tr>
-      <tr>
-        <td>Observaciones:</td>
-        <td colspan="2"><textarea [disabled]="disabled" style="width:100%;" class="text-area" rows="5" cols="25" [(ngModel)]="activity.notes"></textarea></td>
-      </tr>
-    </table>
-  </div>
-  <div>
-    <table class="form">
-      <tr>
-        <td class="form-title form-title-text">2) Asignación de tareas</td>
-      </tr>
-      <tr>
-        <table class="form grid">
-          <thead>
-            <tr>
-              <th style="width:5%">#</th>
-              <th style="width:55%">Tramite/Documentos</th>
-              <th style="width:40%">Asignado a</th>
-            </tr>
-          </thead>
-          <tbody>
-            <tr *ngFor="let task of taksList; let i = index;">
-              <td style="width:5%">{{i +1}}</td>
-              <td style="width:55%">{{task.name}}</td>
-              <td style="width:40%"><select class="select-box"><option value="">( Seleccionar )</option>
-                <option *ngFor="let taskManager of taskManagersList" 
-                        [value]="taskManager.uid">{{taskManager.name}}</option>            
-                </select></td>
-            </tr>
-          </tbody>
-        </table>
-      </tr>
-    </table>
-  </div>
-  <div class="form-footer border-top">
-    <table class="form">
-      <tr>
-        <td class="text-align-right ">
-          <button class="btn " (click)="onClickCancel()">Cancelar</button>
-        </td>
-        <td class="text-align-right ">
-          <button class="btn " (click)="onClickAddActivity()">Iniciar la actividad</button>
-        </td>
-      </tr>
-    </table>
-  </div>
-</div>
