--- src/app/projects/activity-editor/project-start-activity.component.scss
@@ -1,21 +1,0 @@
-@import "../../../styles/editor-window-style";
-@import "../../../styles/form";
-@import "../../../styles/grid";
-.form {
-  width: 640px;
-}
-
-:host {
-  position: relative;
-  top: -25%;
-  left: 20%;
-}
-
-.editor-container {
-  height: 630px;
-  width: 700px;
-}
-
-table.grid tbody {
-  height: 170px;
-}
