+++ src/app/projects/activity-editor/tabs/project-assign-tasks.component.html
@@ -1,0 +1,46 @@
+<br>
+<div>
+  <table class="form">
+    <tr>
+      <td class="form-title form-title-text">2) Asignación de tareas</td>
+    </tr>
+    <tr>
+      <table class="form grid">
+        <thead>
+          <tr>
+            <th style="width:5%">#</th>
+            <th style="width:53%">Tramite/Documentos</th>
+            <th style="width:30%">Asignado a</th>
+            <th style="width:12%">.</th>
+          </tr>
+        </thead>
+        <tbody>
+          <tr *ngFor="let task of taksList; let i = index;">
+            <td style="width:5%">{{i +1}}</td>
+            <td style="width:53%">{{task.name}}</td>
+            <td style="width:30%"><select class="select-box"><option value="">( Seleccionar )</option>
+                <option *ngFor="let taskManager of taskManagersList" 
+                        [value]="taskManager.uid">{{taskManager.name}}</option>            
+                </select>
+            </td>
+            <td style="width:12%" class="text-align-left"><button class="btn">+Asignar</button></td>
+          </tr>
+        </tbody>
+      </table>
+    </tr>
+  </table>
+</div>
+<div class="form-footer border-top">
+  <table class="form">
+    <tr>
+      <!--
+      <td class="text-align-right ">
+        <button class="btn " (click)="onClickCancel()">Cancelar</button>
+      </td>
+      <td class="text-align-right ">
+        <button class="btn " (click)="onClickAddActivity()">Iniciar la actividad</button>
+      </td>-->
+    </tr>
+  </table>
+
+</div>
