+++ src/app/projects/activity-editor/tabs/project-assign-tasks.component.scss
@@ -1,0 +1,9 @@
+@import "../../../../styles/form";
+@import "../../../../styles/grid";
+.form {
+  width: 640px;
+}
+
+table.grid tbody {
+  height: 300px;
+}
