+++ src/app/projects/activity-editor/tabs/project-start-activity.component.scss
@@ -1,0 +1,5 @@
+@import "../../../../styles/form";
+.form {
+  width: 640px;
+  overflow-y: auto;
+}
